WIP: Refactoring sprite editor out
